// src/types/ChickenData.ts
export interface ChickenData {
    id: number;
    manufacturer: string;
    product_name: string;
    calories: number;
    protein: number;
    fat: number;
    carbohydrate: number;
    sugars: number;
    sodium: number;
    cholesterol: number;
    saturated_fat: number;
    weight: number;
    brand: string;
}